requirements.txt
pytorch >= 0.4.1
pytorch_pretrained_bert >= 0.6.1
allennlp=0.8.2
torchnlp
fastNLP
